+++ example/src/App.tsx
@@ -9,7 +9,8 @@ import DeepLink from "./Examples/DeepLink"
 import OnEvent from "./Examples/OnEvent"
 import Blended from "./Examples/Blended"
 import Login from "./Examples/Login"
-import Player from "./Examples/Player"
+import Async from "./Examples/Async"
+import EventLevels from "./Examples/EventLevels"
 import Editor from "./Editor"
 
 export interface Props {}
@@ -19,7 +20,7 @@ const App: React.FC<Props> = ({ children }) => {
     <div>
       <h1>State Designer</h1>
       <hr />
-      <Player />
+      <EventLevels />
       {/* <Login /> */}
       {/* <Chain />
       <PlayPauseStop /> */}